meta: Make massive restructuring changes to accomodate new model
Previously, the plan was to have Nginx running outside Docker and
forward to the sites, which would be running as Docker containers and
publishing their ports on the host. We are switching away from the
model.

Instead, we are having Nginx run inside a Docker container and forward
requests to the services over an internal network. We are also making
changes in the current testing "production" environment pertaining to
this that require modifications here.