orchestrator: Completely refactor status websocket stream
Moved log-following code into a separate class and made it use a
subprocess instead of the Python Docker API.

The Python Docker API is great, but it's not asyncio-friendly. There is
aiodocker on PyPI, but it hasn't made a 1.0 release and I don't want to
make our Docker client code into even more of a mess.