manager: Remove Action.equivalent_command
The original intention was for this field to store shell commands that
could be run manually if an action failed, to try to recover and perform
whatever action was required. However, this ran into difficulties.

As Director 4.0 took shape, it became clear that most -- in fact, nearly
all -- of the Actions do not translate directly to shell commands, at
least not anything sane. To implement equivalent_command properly, it
would be necessary to write large numbers of management commands to
perform certain actions manually. That would be excessively wasteful.

In addition, nearly all the Actions are now idempotent, and the
`fix_site` operation is capable of recovering a site from nearly any
conceivable state. Hence, there is no need for the equivalent_command
field.