meta: Remove "update password" database helpers
create_database() sets the password, and it also makes sure that the
database is in the proper state. It works fine in update_password()'s
place.

In fact, because of a typo the update-password endpoint on the
orchestrator has been accidentally calling create_database() instead of
update_password() with no issues!