meta: Add concept of site 'availability'
Uses the same model as Domain.status. Credit goes to @theo-o for the
name (and the original concept of Domain.status that this is based off
of).

Essentially, a site's availability can be one of the three values:
- Enabled: Normal functionality.
- Not served: The site is not served publicly, but otherwise it
  functions normally.
- Disabled: The site is not served publicly, and only superusers can
  view/edit site information through the web interface. (Note that for
  non-superusers who normally have access to the site, the site still
  appears on the home page. However, attempting to visit the "site
  information" page will give a descriptive error message, and visiting
  any other page related to that viewing/editing information regarding
  that site will give a 404.)

The "not served"/"disabled" availabilities are implemented by 1) scaling
the Docker service to 0 replicas and 2) editing the Nginx config to
return an error page.

Note that the large number of changes to the views/consumers was
necessary to enforce the new permission restrictions.