docs: Update docs on number of Celery workers
There are no longer major concerns that concurrent modification of
different sites will cause issues.

For example, the type of problem described in this old documentation
(renaming a site while a site with that name is being created) is not an
issue because site IDs are used instead of site names whenever possible.
(In addition, site names are required to be unique at the database
level.)