orchestrator: Pull latest image before launching terminal
This ensures that when the user launches a terminal, it always uses the
latest built image. This may not be the case if the appserver on which
the terminal is launched has not previously pulled the latest image.

This issue was discovered through testing in production. It didn't occur
in development because the testing environment had 2 appservers instead
of 5, so it was less likely to occur.