orchestrator: Fix occasional errors when initially launching terminal
If a user tries to open a terminal but it times out, for the next few
minutes they may immediately see "Disconnected" every time they reload.
This is because the container is still starting, so it doesn't show up
when listing images, but trying to re-create it fails because the name
is in use.

This commit changes the behavior to retry in that case.